feat: user register && refactor some implementation

+++ src/main/java/cn/edu/sdtbu/model/entity/BaseEntity.java
@@ -2,6 +2,7 @@ package cn.edu.sdtbu.model.entity;
 
 import cn.edu.sdtbu.model.properties.Const;
 import cn.edu.sdtbu.util.TimeUtil;
+import com.google.common.primitives.UnsignedLong;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
@@ -23,7 +24,7 @@ import java.sql.Timestamp;
 public class BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column
     private Timestamp createAt;
