+++ src/main/java/cn/edu/sdtbu/repository/UserRepository.java
@@ -2,6 +2,7 @@ package cn.edu.sdtbu.repository;
 
 import cn.edu.sdtbu.model.entity.UserEntity;
 import cn.edu.sdtbu.repository.base.BaseRepository;
+import com.google.common.primitives.UnsignedLong;
 
 /**
  * user repository
@@ -9,5 +10,5 @@ import cn.edu.sdtbu.repository.base.BaseRepository;
  * @version 1.0
  * @date 2020-4-6 20:56
  */
-public interface UserRepository extends BaseRepository<UserEntity, Integer> {
+public interface UserRepository extends BaseRepository<UserEntity, Long> {
 }
